feat: add optional name to import project endpoint
